@@ -18,13 +18,14 @@ import (
 )
 
 type configOpts struct {
-	rsync    bool
-	wait     bool
-	waitTime time.Duration
-	list     bool
-	file     string
-	set      map[string]string
-	key      string
+	rsync           bool
+	includeRestores bool
+	wait            bool
+	waitTime        time.Duration
+	list            bool
+	file            string
+	set             map[string]string
+	key             string
 }
 
 type confKV struct {
@@ -76,8 +77,15 @@ func runConfig(
 			}
 		}
 		if rsnc {
-			if _, err := pbm.SyncFromStorage(ctx); err != nil {
+			cid, err := pbm.SyncFromStorage(ctx, false)
+			if err != nil {
 				return nil, errors.Wrap(err, "resync")
+			}
+
+			if c.wait {
+				if err := waitForResyncWithTimeout(ctx, pbm, cid, c.waitTime); err != nil {
+					return nil, err
+				}
 			}
 		}
 		return o, nil
@@ -91,7 +99,7 @@ func runConfig(
 		}
 		return confKV{c.key, fmt.Sprint(k)}, nil
 	case c.rsync:
-		cid, err := pbm.SyncFromStorage(ctx)
+		cid, err := pbm.SyncFromStorage(ctx, c.includeRestores)
 		if err != nil {
 			return nil, errors.Wrap(err, "resync")
 		}
@@ -100,19 +108,8 @@ func runConfig(
 			return outMsg{"Storage resync started"}, nil
 		}
 
-		if c.waitTime > time.Second {
-			var cancel context.CancelFunc
-			ctx, cancel = context.WithTimeout(ctx, c.waitTime)
-			defer cancel()
-		}
-
-		err = sdk.WaitForResync(ctx, pbm, cid)
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				err = errWaitTimeout
-			}
-
-			return nil, errors.Wrapf(err, "waiting for resync [opid %q]", cid)
+		if err := waitForResyncWithTimeout(ctx, pbm, cid, c.waitTime); err != nil {
+			return nil, err
 		}
 
 		return outMsg{"Storage resync finished"}, nil
@@ -143,8 +140,15 @@ func runConfig(
 
 		// resync storage only if Storage options have changed
 		if !reflect.DeepEqual(newCfg.Storage, oldCfg.Storage) {
-			if _, err := pbm.SyncFromStorage(ctx); err != nil {
+			cid, err := pbm.SyncFromStorage(ctx, false)
+			if err != nil {
 				return nil, errors.Wrap(err, "resync")
+			}
+
+			if c.wait {
+				if err := waitForResyncWithTimeout(ctx, pbm, cid, c.waitTime); err != nil {
+					return nil, err
+				}
 			}
 		}
 
